Refactored command packages to fit RobotTemplate
