+++ src/frc/team3238/commands/extender/Withdraw.java
@@ -1,20 +1,14 @@
-package frc.team3238.commands;
+package frc.team3238.commands.extender;
 
-import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.command.Command;
-import frc.team3238.OI;
 import frc.team3238.Robot;
-import frc.team3238.subsystems.Extender;
 
 public class Withdraw extends Command
 {
-    Extender extender = Robot.extender;
-    OI oi = Robot.oi;
-
     public Withdraw()
     {
         super("Withdraw");
-        requires(extender);
+        requires(Robot.extender);
     }
 
     @Override
@@ -26,9 +20,9 @@ public class Withdraw extends Command
     @Override
     protected void execute()
     {
-        double throttle = oi.getThrottleMult();
+        double throttle = Robot.oi.getThrottleMult();
 
-        extender.setExtend(throttle);
+        Robot.extender.setExtend(throttle);
     }
 
     @Override
@@ -40,7 +34,7 @@ public class Withdraw extends Command
     @Override
     protected void end()
     {
-        extender.stopMotor();
+        Robot.extender.stopMotor();
     }
 
     @Override
